% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/binary_dataset.R
\docType{data}
\name{binary_dataset}
\alias{binary_dataset}
\title{A binary dataframe of datasets used to establish evaluation criteria.}
\format{
A dataframe with 600 rows and 50 variables.
\describe{
\item{"Plant parts or microbial sites" to "Anti-inflammatory"}{Subfeature of nano silver.}
}
}
\usage{
data(binary_dataset)
}
\description{
A binary dataframe of datasets used to establish evaluation criteria.
}
\keyword{datasets}
