% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataset.R
\docType{data}
\name{dataset}
\alias{dataset}
\title{Nanosilver data set.}
\format{
A dataframe with 600 rows and 15 variables.
\describe{
\item{"Synthesis methods" to "Applications"}{Features of nano silver.}
}
}
\usage{
data(dataset)
}
\description{
Nanosilver data set.
}
\keyword{datasets}
