% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/line_radar.R
\name{line_radar}
\alias{line_radar}
\title{Line and Radar Plot of sysAgNPs score}
\usage{
line_radar(sysAgNPs_score, num_plots, output_directory)
}
\arguments{
\item{sysAgNPs_score}{A dataframe containing four columns of numeric data.}

\item{num_plots}{The range of the graph to be output and saved can be a vector or a single value.}

\item{output_directory}{Directory where the plots will be saved.}
}
\value{
A \code{ggplot} object.
}
\description{
Line and Radar Plot of sysAgNPs score
}
\examples{
sysAgNPs_line_radar <- line_radar(sysAgNPs_score, 10, "sysAgNPs_line_radar/")
}
