% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sys_CalculateAxisPath.R
\name{sys_CalculateAxisPath}
\alias{sys_CalculateAxisPath}
\title{Calculate Axis Path.
This function is derived from the 'ggradar' package.\url{https://github.com/ricardo-bion/ggradar/}.
Calculates x-y coordinates for a set of radial axes (one per variable being plotted in radar plot)}
\usage{
sys_CalculateAxisPath(var.names, min, max)
}
\arguments{
\item{var.names}{list of variables to be plotted on radar plot}

\item{min}{MININUM value required for the plotted axes (same value will be applied to all axes)}

\item{max}{MAXIMUM value required for the plotted axes (same value will be applied to all axes)}
}
\value{
a dataframe of the calculated axis paths
}
\description{
Calculate Axis Path.
This function is derived from the 'ggradar' package.\url{https://github.com/ricardo-bion/ggradar/}.
Calculates x-y coordinates for a set of radial axes (one per variable being plotted in radar plot)
}
