% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sys_CE.R
\name{sys_CE}
\alias{sys_CE}
\title{Calculate the Combination Entropy}
\usage{
sys_CE(data, dataset)
}
\arguments{
\item{data}{A dataframe that contains experimental data.}

\item{dataset}{The dataset used to to calculate the ratio of the number of reporting a certain feature in the AgNPs dataset to the total number of samples.}
}
\value{
A dataframe including: 1. the ratio of the number of reporting a certain feature in the AgNPs dataset to the total number of samples; 2. pc:the probability of the feature combination occurring; 3. Hi:the probability of surprising level of the presence and absence of feature combinations to measure the uncertainty of feature combination.
}
\description{
Calculate the average value of the probability of surprising level of the  presence and absence of a particular category within the specific category to measure the average uncertainty of feature categories.
}
\examples{
data(dataset)
users_data <- dataset
CE <- sys_CE(users_data, dataset)
}
