% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sys_DE.R
\name{sys_DE}
\alias{sys_DE}
\title{Calculate the Distribution Entropy}
\usage{
sys_DE(data)
}
\arguments{
\item{data}{A dataframe that contains experimental data.}
}
\value{
A dataframe including 1. the number of feature in a certain category; 2. the total number of features in the sample; 3.  the average value to measure the average uncertainty of feature categories
}
\description{
Measure the distribution variability of the presence and absence of feature categories.
}
\examples{
DE = sys_DE(users_data)

}
