% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sys_discretize.R
\name{sys_discretize}
\alias{sys_discretize}
\title{Convert categorical variables into discrete variables}
\usage{
sys_discretize(dataset, vars_to_discretize = NULL)
}
\arguments{
\item{dataset}{A dataframe of dataset. Datasets used to establish evaluation criteria.}

\item{vars_to_discretize}{Variables or columns to be discretized. Default is NULL.}
}
\value{
A binary dataframe.
}
\description{
Convert categorical variables into discrete variables
}
\examples{
data(dataset)
dis_data <- sys_discretize(dataset, c("Shape", "pH"))
}
