% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sys_funcCircleCoords.R
\name{sys_funcCircleCoords}
\alias{sys_funcCircleCoords}
\title{Generate circle coordinates.
This function is derived from the 'ggradar' package.\url{https://github.com/ricardo-bion/ggradar/}.
Generate coordinates to draw a circle.}
\source{
Adapted from Joran's response to \url{https://stackoverflow.com/questions/6862742/draw-a-circle-with-ggplot2/}.
}
\usage{
sys_funcCircleCoords(center = c(0, 0), r = 1, npoints = 100)
}
\arguments{
\item{center}{coordinate for centroid}

\item{r}{radius}

\item{npoints}{number of coordinates to generate}
}
\value{
a dataframe
}
\description{
Generate circle coordinates.
This function is derived from the 'ggradar' package.\url{https://github.com/ricardo-bion/ggradar/}.
Generate coordinates to draw a circle.
}
