% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sys_generate_color_values.R
\name{sys_generate_color_values}
\alias{sys_generate_color_values}
\title{Generate Dynamic Color Values.
This function is derived from the 'ggradar' package.\url{https://github.com/ricardo-bion/ggradar/}.
This function dynamically generates a vector of color values based on the
number of groups. It uses RColorBrewer for smaller sets of groups and
generates a gradient for larger sets.}
\usage{
sys_generate_color_values(num_groups)
}
\arguments{
\item{num_groups}{The number of groups for which to generate color values.}
}
\value{
A character vector of color values.
}
\description{
Generate Dynamic Color Values.
This function is derived from the 'ggradar' package.\url{https://github.com/ricardo-bion/ggradar/}.
This function dynamically generates a vector of color values based on the
number of groups. It uses RColorBrewer for smaller sets of groups and
generates a gradient for larger sets.
}
\examples{
sys_generate_color_values(5)
sys_generate_color_values(20)
}
