% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sys_PE.R
\name{sys_PE}
\alias{sys_PE}
\title{Calculate the Proclivity Entropy}
\usage{
sys_PE(data)
}
\arguments{
\item{data}{A dataframe that contains experimental data.}
}
\value{
A dataframe including 1. the number of feature in a certain category; 2. the total number of features in the sample; 3. the expected value to measure the average description level across different feature categories.
}
\description{
Measure the preference of feature categories.
}
\examples{
PE = sys_PE(users_data)
}
