% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sys_TS.R
\name{sys_TS}
\alias{sys_TS}
\title{Calculate the Transition Scores}
\usage{
sys_TS(data, dataset, n_iter, vars_to_discretize)
}
\arguments{
\item{data}{A dataframe that contains experimental data.}

\item{dataset}{A binary dataframe. Datasets used to establish evaluation criteria.}

\item{n_iter}{The number of iterations to reach the steady state.}

\item{vars_to_discretize}{Variables or columns to be discretized. Default is NULL.}
}
\value{
A dataframe that contains sysAgNPs scores.
}
\description{
Calculate the Transition Scores
}
