% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sysAgNPs_score.R
\docType{data}
\name{sysAgNPs_score}
\alias{sysAgNPs_score}
\title{sysAgNPs package application results in four evaluation scores.}
\format{
A dataframe with 600 rows and 4 variables.
\describe{
\item{DE to TS}{Four evaluation scores.}
}
}
\usage{
data(sysAgNPs_score)
}
\description{
sysAgNPs package application results in four evaluation scores.
}
\keyword{datasets}
