% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tran_matrix.R
\docType{data}
\name{tran_matrix}
\alias{tran_matrix}
\title{A transfer probability matrix.}
\format{
A matrix with 50 rows and 50 columns.
\describe{
\item{V1 to V50}{Subfeature of nano silver.}
}
}
\usage{
data(tran_matrix)
}
\description{
A transfer probability matrix.
}
\keyword{datasets}
