% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sys_tran.R
\name{sys_tran}
\alias{sys_tran}
\title{Calculate transition probability matrix}
\usage{
sys_tran(binary_dataset)
}
\arguments{
\item{binary_dataset}{A binary dataframe of datasets used to establish Transition Scores criteria.}
}
\value{
A transfer probability matrix.
}
\description{
Calculate transition probability matrix
}
\examples{
tran_matrix <- sys_tran(binary_dataset)
}
